server/requirements.txt
fastapi
uvicorn
pydantic
passlib
bcrypt==4.0.1
sqlalchemy
databases[aiosqlite]
python-jose[cryptography]
python-multipart
requests
pytest
httpx
python-dotenv
pydantic-settings
websockets
alembic